OneDrive/Proyectos/PCAF Pro/backend/requirements.txt
# backend/requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
PyJWT==2.1.0
python-dotenv==0.19.0